adding sandbox folder
working on basic functionality in this folder before adding it to the
main program. Front end of the question submit is working: it currently
just needs to store the object correctly in the nbd. Once in the nbd,
class ReviewQuestion can be modified to load a dynamic question based
on a questionID. The Sandbox folder is designed to run as it’s own app
engine project